moved interfaces and enums into submodule
