--- src/roles/role.enum.ts
@@ -1,4 +1,0 @@
-export enum Role {
-    Employee = 'employee',
-    Admin = 'admin',
-}
